fix: working healthchecks and delayed shutdown
